requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-CORS
psycopg2-binary
python-dotenv
bcrypt
